+++ frontend/android/app/src/main/java/com/dorukefty/pcmanager/MainActivity.java
@@ -1,0 +1,5 @@
+package com.dorukefty.pcmanager;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}